Finished implementing my mockup to_digit()

+++ src/main.rs
@@ -26,7 +26,7 @@ fn main() {
         "1" => println!("Binary to decimal."),
         "2" => {
             println!("Binary to hexadecimal.");
-            let mut binary_number: Vec<u32> = conversions::get_input();
+            let mut binary_number: Vec<u8> = conversions::get_input();
             let hex_number: Vec<String> = conversions::binary_hex(&mut binary_number);
             println!("Your number in hexadecimal is: ");
             for i in hex_number {
